Split resolvers out
